requirements.txt
fastapi[standard]
web3
pyjwt
sqlmodule
